Create a model describing SourceCode
Because the current situation concerning code generators was found to be hard
to maintain and was inviting for code copy which is pure evil,
an attempt was made to describe source with a model to formalize the code
generation process a bit. The idea is not to use something like format strings
in the language specific generators like this:
'''
void «type.toFirstLower»_exit(«scHandleDecl»)
{
	«exitSequence.code»
}
'''
but rather to create a function programmatically:
new Function()
	.name(...)
	.return(VOID)
	.addParam(...)
	.addBody(...)
so that inheritance becomes actually possible because this code might only need
to be written once and reused for all languages, but with different language
definitions in the background.

This is an experiment and will probably change a lot during work.